requirements.txt
PySide6==6.7.2
pyobjc==10.3.1
websockets==12.0
